refactor to abstract Decimals type

+++ app/ante/testutils/testutils.go
@@ -109,7 +109,7 @@ func (suite *AnteTestSuite) SetupTest() {
 	configurator.ResetTestChainConfig()
 	err := configurator.
 		WithChainConfig(chainConfig).
-		WithEVMCoinInfo(denom, decimals).
+		WithEVMCoinInfo(denom, int64(decimals)).
 		Configure()
 	suite.Require().NoError(err)
